changed prices to BigDecimal / new field in Reservation / test fixes wip / statistics fix / new column / added temp validator for reservation price

+++ src/main/java/CPSF/com/demo/controller/CamperPlaceController.java
@@ -7,6 +7,7 @@ import CPSF.com.demo.request.CamperPlaceRequest;
 import CPSF.com.demo.service.CamperPlaceService;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
@@ -22,7 +23,7 @@ public class CamperPlaceController {
 
     @PostMapping("/create")
     public void createCamperPlace(@RequestBody CamperPlaceRequest request) {
-        camperPlaceService.create(request.type(), request.price());
+        camperPlaceService.create(request.type(), BigDecimal.valueOf(request.price()));
     }
 
     @GetMapping("/getAll")
