+++ src/test/java/CPSF/com/demo/reservationCalculatorTest/ReservationCalculatorTest.java
@@ -3,6 +3,8 @@ package CPSF.com.demo.reservationCalculatorTest;
 import CPSF.com.demo.util.ReservationCalculator;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
@@ -12,18 +14,18 @@ public class ReservationCalculatorTest {
 
 	@Test
 	public void shouldReturnCorrectCostWithNoDiscount() {
-		double camperPlacePrice = 100;
-		double expected = 300;
-		double actual = reservationCalculator.calculateFinalReservationCost(3, camperPlacePrice);
+		BigDecimal camperPlacePrice = BigDecimal.valueOf(100);
+		BigDecimal expected = BigDecimal.valueOf(300);
+		BigDecimal actual = reservationCalculator.calculateFinalReservationCost(3, camperPlacePrice);
 
 		assertEquals(expected, actual);
 	}
 
 	@Test
 	public void shouldReturnCorrectCostWithDiscount() {
-		double camperPlacePrice = 100;
-		double expected = 930;
-		double actual = reservationCalculator.calculateFinalReservationCost(10, camperPlacePrice);
+		BigDecimal camperPlacePrice = BigDecimal.valueOf(100);
+		BigDecimal expected = BigDecimal.valueOf(930);
+		BigDecimal actual = reservationCalculator.calculateFinalReservationCost(10, camperPlacePrice);
 
 		assertEquals(expected, actual);
 	}
